+++ AESound/AppDelegate.swift
@@ -1,5 +1,5 @@
 /**
- *  https://github.com/tadija/AESystemSound
+ *  https://github.com/tadija/AESound
  *  Copyright (c) Marko Tadić 2018
  *  Licensed under the MIT license. See LICENSE file.
  */