+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## Version 0.3.0
+
+- Renamed project from `AESystemSound` -> `AESound`
+- Added both `AudioToolbox` and `AVFoundation` variants for playing sound
+- Added search and framework toggle to demo project
+- Other improvements and refactoring
+
 ## Version 0.2.0
 
-- New `AESystemSound.Path` enum
+- New enum with all system sound paths (iOS 11.4.1)
 
 ## Version 0.1.0